增加自动清理mysql存储空间

+++ beecomb-common/src/main/java/io/github/icodegarden/beecomb/common/metrics/job/JobsMetricsOverload.java
@@ -114,7 +114,7 @@ public class JobsMetricsOverload implements MetricsOverload {
 		return metrics;
 	}
 
-	public void enableScheduleFlushMetrics(int scheduleMillis) {
+	public void enableScheduleFlushMetrics(long scheduleMillis) {
 		defaultMetricsOverload.enableScheduleFlushMetrics(scheduleMillis);
 	}
